requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
channels
Pillow
whitenoise
